requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
fastembed
faiss-cpu
python-dotenv
ragas
langsmith
google-api-core